extract viewport targeting
